pyproject.toml
[project]
name = "virtuoso-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aioice==0.10.1",
    "aiortc==1.14.0",
    "alembic==1.17.1",
    "annotated-doc==0.0.3",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "av==16.0.1",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "click==8.3.0",
    "colorama==0.4.6",
    "cryptography==46.0.3",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi==0.120.1",
    "fastapi-cli==0.0.14",
    "fastapi-cloud-cli==0.3.1",
    "google-crc32c==1.7.1",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "ifaddr==0.2.0",
    "iniconfig==2.3.0",
    "jinja2==3.1.6",
    "mako==1.3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "packaging==25.0",
    "pluggy==1.6.0",
    "psycopg2==2.9.11",
    "pwdlib==0.3.0",
    "pycparser==2.23",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pydantic-settings==2.11.0",
    "pyee==13.0.0",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pylibsrtp==1.0.0",
    "pyopenssl==25.3.0",
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.3",
    "rich==14.2.0",
    "rich-toolkit==0.15.1",
    "rignore==0.7.1",
    "sentry-sdk==2.42.1",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.44",
    "sqlmodel==0.0.27",
    "starlette==0.49.1",
    "typer==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "watchfiles==1.1.1",
    "websockets==15.0.1",
]
